requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
pydantic==2.10.3
google-generativeai==0.8.3
python-multipart==0.0.20
requests==2.32.3
